move subsystem health to dashboard
